Broadcom Android Release 7.0 Phase 16.3 (eng 1) code drop

+++ nexus/platforms/common/src/nexus_platform_debug_log.c
@@ -1,51 +1,40 @@
-/***************************************************************************
-*     (c)2011-2014 Broadcom Corporation
-*
-*  This program is the proprietary software of Broadcom Corporation and/or its licensors,
-*  and may only be used, duplicated, modified or distributed pursuant to the terms and
-*  conditions of a separate, written license agreement executed between you and Broadcom
-*  (an "Authorized License").  Except as set forth in an Authorized License, Broadcom grants
-*  no license (express or implied), right to use, or waiver of any kind with respect to the
-*  Software, and Broadcom expressly reserves all rights in and to the Software and all
-*  intellectual property rights therein.  IF YOU HAVE NO AUTHORIZED LICENSE, THEN YOU
-*  HAVE NO RIGHT TO USE THIS SOFTWARE IN ANY WAY, AND SHOULD IMMEDIATELY
-*  NOTIFY BROADCOM AND DISCONTINUE ALL USE OF THE SOFTWARE.
-*
-*  Except as expressly set forth in the Authorized License,
-*
-*  1.     This program, including its structure, sequence and organization, constitutes the valuable trade
-*  secrets of Broadcom, and you shall use all reasonable efforts to protect the confidentiality thereof,
-*  and to use this information only in connection with your use of Broadcom integrated circuit products.
-*
-*  2.     TO THE MAXIMUM EXTENT PERMITTED BY LAW, THE SOFTWARE IS PROVIDED "AS IS"
-*  AND WITH ALL FAULTS AND BROADCOM MAKES NO PROMISES, REPRESENTATIONS OR
-*  WARRANTIES, EITHER EXPRESS, IMPLIED, STATUTORY, OR OTHERWISE, WITH RESPECT TO
-*  THE SOFTWARE.  BROADCOM SPECIFICALLY DISCLAIMS ANY AND ALL IMPLIED WARRANTIES
-*  OF TITLE, MERCHANTABILITY, NONINFRINGEMENT, FITNESS FOR A PARTICULAR PURPOSE,
-*  LACK OF VIRUSES, ACCURACY OR COMPLETENESS, QUIET ENJOYMENT, QUIET POSSESSION
-*  OR CORRESPONDENCE TO DESCRIPTION. YOU ASSUME THE ENTIRE RISK ARISING OUT OF
-*  USE OR PERFORMANCE OF THE SOFTWARE.
-*
-*  3.     TO THE MAXIMUM EXTENT PERMITTED BY LAW, IN NO EVENT SHALL BROADCOM OR ITS
-*  LICENSORS BE LIABLE FOR (i) CONSEQUENTIAL, INCIDENTAL, SPECIAL, INDIRECT, OR
-*  EXEMPLARY DAMAGES WHATSOEVER ARISING OUT OF OR IN ANY WAY RELATING TO YOUR
-*  USE OF OR INABILITY TO USE THE SOFTWARE EVEN IF BROADCOM HAS BEEN ADVISED OF
-*  THE POSSIBILITY OF SUCH DAMAGES; OR (ii) ANY AMOUNT IN EXCESS OF THE AMOUNT
-*  ACTUALLY PAID FOR THE SOFTWARE ITSELF OR U.S. $1, WHICHEVER IS GREATER. THESE
-*  LIMITATIONS SHALL APPLY NOTWITHSTANDING ANY FAILURE OF ESSENTIAL PURPOSE OF
-*  ANY LIMITED REMEDY.
-*
-* $brcm_Workfile: $
-* $brcm_Revision: $
-* $brcm_Date: $
-*
-* API Description:
-*
-* Revision History:
-*
-* $brcm_Log: $
-*
-***************************************************************************/
+/******************************************************************************
+ *  Broadcom Proprietary and Confidential. (c)2016 Broadcom. All rights reserved.
+ *
+ *  This program is the proprietary software of Broadcom and/or its licensors,
+ *  and may only be used, duplicated, modified or distributed pursuant to the terms and
+ *  conditions of a separate, written license agreement executed between you and Broadcom
+ *  (an "Authorized License").  Except as set forth in an Authorized License, Broadcom grants
+ *  no license (express or implied), right to use, or waiver of any kind with respect to the
+ *  Software, and Broadcom expressly reserves all rights in and to the Software and all
+ *  intellectual property rights therein.  IF YOU HAVE NO AUTHORIZED LICENSE, THEN YOU
+ *  HAVE NO RIGHT TO USE THIS SOFTWARE IN ANY WAY, AND SHOULD IMMEDIATELY
+ *  NOTIFY BROADCOM AND DISCONTINUE ALL USE OF THE SOFTWARE.
+ *
+ *  Except as expressly set forth in the Authorized License,
+ *
+ *  1.     This program, including its structure, sequence and organization, constitutes the valuable trade
+ *  secrets of Broadcom, and you shall use all reasonable efforts to protect the confidentiality thereof,
+ *  and to use this information only in connection with your use of Broadcom integrated circuit products.
+ *
+ *  2.     TO THE MAXIMUM EXTENT PERMITTED BY LAW, THE SOFTWARE IS PROVIDED "AS IS"
+ *  AND WITH ALL FAULTS AND BROADCOM MAKES NO PROMISES, REPRESENTATIONS OR
+ *  WARRANTIES, EITHER EXPRESS, IMPLIED, STATUTORY, OR OTHERWISE, WITH RESPECT TO
+ *  THE SOFTWARE.  BROADCOM SPECIFICALLY DISCLAIMS ANY AND ALL IMPLIED WARRANTIES
+ *  OF TITLE, MERCHANTABILITY, NONINFRINGEMENT, FITNESS FOR A PARTICULAR PURPOSE,
+ *  LACK OF VIRUSES, ACCURACY OR COMPLETENESS, QUIET ENJOYMENT, QUIET POSSESSION
+ *  OR CORRESPONDENCE TO DESCRIPTION. YOU ASSUME THE ENTIRE RISK ARISING OUT OF
+ *  USE OR PERFORMANCE OF THE SOFTWARE.
+ *
+ *  3.     TO THE MAXIMUM EXTENT PERMITTED BY LAW, IN NO EVENT SHALL BROADCOM OR ITS
+ *  LICENSORS BE LIABLE FOR (i) CONSEQUENTIAL, INCIDENTAL, SPECIAL, INDIRECT, OR
+ *  EXEMPLARY DAMAGES WHATSOEVER ARISING OUT OF OR IN ANY WAY RELATING TO YOUR
+ *  USE OF OR INABILITY TO USE THE SOFTWARE EVEN IF BROADCOM HAS BEEN ADVISED OF
+ *  THE POSSIBILITY OF SUCH DAMAGES; OR (ii) ANY AMOUNT IN EXCESS OF THE AMOUNT
+ *  ACTUALLY PAID FOR THE SOFTWARE ITSELF OR U.S. $1, WHICHEVER IS GREATER. THESE
+ *  LIMITATIONS SHALL APPLY NOTWITHSTANDING ANY FAILURE OF ESSENTIAL PURPOSE OF
+ *  ANY LIMITED REMEDY.
+ ******************************************************************************/
 #include "nexus_platform.h"
 #include "nexus_platform_priv.h"
 #include "nexus_platform_debug_log.h"
@@ -66,12 +55,16 @@ void NEXUS_Platform_P_DebugLog_Init(NEXUS_Platform_P_DebugLog *debugLog, const c
     BDBG_Fifo_CreateSettings logSettings;
     const char *debug_log_size;
     const char *logger;
-    const char *fname;
+    const char *fname=NULL;
+    int nread;
     int urc;
     BERR_Code rc;
     struct stat st;
     int pipesfd[2];
     char fd[4];
+
+    BKNI_Memset(debugLog, 0, sizeof(*debugLog));
+    debugLog->fd = -1;
 
     /* Allow override of logger */
     logger = NEXUS_GetEnv("nexus_logger");
@@ -90,7 +83,9 @@ void NEXUS_Platform_P_DebugLog_Init(NEXUS_Platform_P_DebugLog *debugLog, const c
     } else if ( strlen(fname) >= sizeof(debugLog->fname) ) {
         BDBG_WRN(("Logger file path too long - reverting to default"));
         fname = DEFAULT_LOGGER_FILE;
-    } else {
+    } else if ( !strcmp(fname, "disabled") ) {
+        BDBG_WRN(("Logger file disabled"));
+    }  else {
         BDBG_WRN(("Logger data file overridden to %s", fname));
     }
 
@@ -99,6 +94,7 @@ void NEXUS_Platform_P_DebugLog_Init(NEXUS_Platform_P_DebugLog *debugLog, const c
     if(urc<0) {goto err_logger;}
 
     b_strncpy(debugLog->fname, fname, sizeof(debugLog->fname));
+
     BDBG_Fifo_GetDefaultCreateSettings(&logSettings);
     BDBG_Log_GetElementSize(&logSettings.elementSize);
     logSettings.nelements = 1024;
@@ -113,30 +109,34 @@ void NEXUS_Platform_P_DebugLog_Init(NEXUS_Platform_P_DebugLog *debugLog, const c
         goto err_elements;
     }
     debugLog->logSize = logSettings.nelements;
-    unlink(debugLog->fname);
-    debugLog->fd = open(fname, O_RDWR|O_CREAT|O_TRUNC, 0666);
-    if(debugLog->fd<0) { (void)BERR_TRACE(NEXUS_OS_ERROR);goto err_fd;}
-    urc = fcntl(debugLog->fd, F_GETFL);
-    if (urc != -1) {
-        urc = fcntl(debugLog->fd, F_SETFL, urc | FD_CLOEXEC);
+
+    if ( strcmp(fname, "disabled") ) {
+        unlink(debugLog->fname);
+        debugLog->fd = open(fname, O_RDWR|O_CREAT|O_TRUNC, 0666);
+        if(debugLog->fd<0) { (void)BERR_TRACE(NEXUS_OS_ERROR);goto err_fd;}
+        urc = fcntl(debugLog->fd, F_GETFL);
+        if (urc != -1) {
+            urc = fcntl(debugLog->fd, F_SETFL, urc | FD_CLOEXEC);
+        }
+        if (urc) BERR_TRACE(urc); /* keep going */
+        debugLog->bufferSize = (logSettings.nelements+1) *  logSettings.elementSize;
+
+        urc = ftruncate(debugLog->fd, debugLog->bufferSize);
+        if(urc<0) { (void)BERR_TRACE(NEXUS_OS_ERROR);goto err_truncate;}
+
+        debugLog->shared = mmap64(0, debugLog->bufferSize, PROT_READ|PROT_WRITE, MAP_SHARED, debugLog->fd, 0);
+        if(debugLog->shared==MAP_FAILED) { (void)BERR_TRACE(NEXUS_OS_ERROR);goto err_mmap;}
+
+        logSettings.buffer = debugLog->shared;
+        logSettings.bufferSize = debugLog->bufferSize;
+         /* coverity[ tainted_data : FALSE ] */
+        rc = BDBG_Fifo_Create(&debugLog->logWriter, &logSettings);
+        if(rc!=BERR_SUCCESS) {(void)BERR_TRACE(rc);goto err_fifo;}
     }
-    if (urc) BERR_TRACE(urc); /* keep going */
-    debugLog->bufferSize = (logSettings.nelements+1) *  logSettings.elementSize;
-
-    urc = ftruncate(debugLog->fd, debugLog->bufferSize);
-    if(urc<0) { (void)BERR_TRACE(NEXUS_OS_ERROR);goto err_truncate;}
-
-    debugLog->shared = mmap64(0, debugLog->bufferSize, PROT_READ|PROT_WRITE, MAP_SHARED, debugLog->fd, 0);
-    if(debugLog->shared==MAP_FAILED) { (void)BERR_TRACE(NEXUS_OS_ERROR);goto err_mmap;}
 
     urc = pipe(pipesfd);
     if(urc<0) { (void)BERR_TRACE(NEXUS_OS_ERROR);goto err_pipe;}
 
-    logSettings.buffer = debugLog->shared;
-    logSettings.bufferSize = debugLog->bufferSize;
-     /* coverity[ tainted_data : FALSE ] */
-    rc = BDBG_Fifo_Create(&debugLog->logWriter, &logSettings);
-    if(rc!=BERR_SUCCESS) {(void)BERR_TRACE(rc);goto err_fifo;}
     debugLog->logger = fork();
     if(debugLog->logger<0) {
         goto err_fork;
@@ -159,24 +159,33 @@ void NEXUS_Platform_P_DebugLog_Init(NEXUS_Platform_P_DebugLog *debugLog, const c
         return;
     }
     close(pipesfd[1]);
-    read(pipesfd[0],fd,1); /* wait for logger to reply */
+    nread = read(pipesfd[0],fd,1); /* wait for logger to reply */
+    if (nread < 0) goto err_fifo;
     close(pipesfd[0]);
-    BDBG_Log_SetFifo(debugLog->logWriter);
+    if ( debugLog->logWriter ) {
+        BDBG_Log_SetFifo(debugLog->logWriter);
+    }
 
     return;
 
 err_fork:
-    BDBG_Fifo_Destroy(debugLog->logWriter);
-    debugLog->logWriter = NULL;
-err_fifo:
     close(pipesfd[0]);
     close(pipesfd[1]);
 err_pipe:
-    munmap(debugLog->shared, debugLog->bufferSize);
+    if ( debugLog->logWriter ) {
+        BDBG_Fifo_Destroy(debugLog->logWriter);
+        debugLog->logWriter = NULL;
+    }
+err_fifo:
+    if ( debugLog->shared ) {
+        munmap(debugLog->shared, debugLog->bufferSize);
+    }
 err_mmap:
 err_truncate:
-    unlink(debugLog->fname);
-    close(debugLog->fd);
+    if ( strcmp(fname, "disabled") ) {
+        unlink(debugLog->fname);
+        close(debugLog->fd);
+    }
 err_fd:
 err_elements:
 err_logger:
@@ -188,20 +197,27 @@ void NEXUS_Platform_P_DebugLog_Uninit(NEXUS_Platform_P_DebugLog *debugLog)
     int status;
     pid_t target;
 
-    if(debugLog->logWriter==NULL) {
-        return;
+    if(debugLog->logWriter) {
+        BKNI_Sleep(10); /* there is no any good way to synchronize with the writer */
+        BDBG_Log_SetFifo(NULL);
     }
-    BKNI_Sleep(10); /* there is no any good way to synchronize with the writer */
-    BDBG_Log_SetFifo(NULL);
-    status = kill(debugLog->logger, SIGUSR1);
-    if (status) {
-        BDBG_ERR(("kill of logger failed with %d", errno));
+    if ( debugLog->logger ) {
+        status = kill(debugLog->logger, SIGUSR1);
+        if (status) {
+            BDBG_ERR(("kill of logger failed with %d", errno));
+        }
+        target = waitpid(debugLog->logger, &status, 0);
+        BDBG_MSG(("%d waitpid:%d, %d", (int)target, (int)debugLog->logger, (int)status));
     }
-    target = waitpid(debugLog->logger, &status, 0);
-    BDBG_MSG(("%d waitpid:%d, %d", (int)target, (int)debugLog->logger, (int)status));
-    BDBG_Fifo_Destroy(debugLog->logWriter);
-    munmap(debugLog->shared, debugLog->bufferSize);
-    close(debugLog->fd);
+    if ( debugLog->logWriter ) {
+        BDBG_Fifo_Destroy(debugLog->logWriter);
+    }
+    if ( debugLog->shared ) {
+        munmap(debugLog->shared, debugLog->bufferSize);
+    }
+    if ( debugLog->fd >= 0 ) {
+        close(debugLog->fd);
+    }
     return;
 }
